Move extractIdFromUrl from api to utils
