+++ src/api/index.ts
@@ -9,13 +9,4 @@ const api = Axios.create({
     headers : {}
 });
 
-export const extractIdFromUrl = (urlType: string, str: string): number => {
-    const idx = str.indexOf(urlType);
-    const match  = /^\/(?<id>\d+)\/?/.exec(str.slice(idx + urlType.length));
-    if (!match || !match.groups) {
-        return -1;
-    }
-    return +match.groups.id;
-};
-
 export default api;
