+++ src/tests/utils.test.ts
@@ -43,4 +43,18 @@ describe('Utils', () => {
             expect(utils.pokemonSorter(pokemons, SortFilter.Default)).toEqual(pokemons);
         });
     });
+
+    describe('extractIdFromUrl', () => {
+        it('Should extract item id from a url correctly', () => {
+            const url = 'https://potato.com/pokemon/3';
+
+            expect(utils.extractIdFromUrl('/pokemon', url)).toBe(3);
+        });
+
+        it('Should return -1 when no id could be extracted', () => {
+            const url = 'https://potato.com/pokemon/test/123';
+
+            expect(utils.extractIdFromUrl('/pokemon', url)).toBe(-1);
+        });
+    });
 });
